# ⚡ Use stack-allocated span for search movegen