--- a/src/Lynx.Benchmark/ParseGame.cs
+++ b/src/Lynx.Benchmark/ParseGame.cs
@@ -371,7 +371,7 @@ public partial class ParseGameBenchmark : BaseBenchmark
 
     public static class ParseGame_ImprovedClass4
     {
-        private static Move[] MovePool = new Move[Constants.MaxNumberOfPossibleMovesInAPosition];
+        private static readonly Move[] _movePool = new Move[Constants.MaxNumberOfPossibleMovesInAPosition];
 
         public const string Id = "position";
 
@@ -411,7 +411,7 @@ public partial class ParseGameBenchmark : BaseBenchmark
                 Span<Range> moves = stackalloc Range[2048]; // Number of potential half-moves provided in the string
                 movesSection.Split(moves, ' ', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
 
-                return new Game(fen, movesSection, moves, MovePool);
+                return new Game(fen, movesSection, moves, _movePool);
             }
             catch (Exception e)
             {
@@ -423,7 +423,7 @@ public partial class ParseGameBenchmark : BaseBenchmark
 
     public static class ParseGame_ImprovedClass5
     {
-        private static Move[] MovePool = new Move[Constants.MaxNumberOfPossibleMovesInAPosition];
+        private static readonly Move[] _movePool = new Move[Constants.MaxNumberOfPossibleMovesInAPosition];
 
         public const string Id = "position";
 
@@ -463,7 +463,7 @@ public partial class ParseGameBenchmark : BaseBenchmark
                 Span<Range> moves = stackalloc Range[(movesSection.Length / 5) + 1]; // Number of potential half-moves provided in the string
                 movesSection.Split(moves, ' ', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
 
-                return new Game(fen, movesSection, moves, MovePool);
+                return new Game(fen, movesSection, moves, _movePool);
             }
             catch (Exception e)
             {