--- a/tests/Lynx.Test/MoveGeneration/GeneralMoveGeneratorTest.cs
+++ b/tests/Lynx.Test/MoveGeneration/GeneralMoveGeneratorTest.cs
@@ -12,10 +12,12 @@ public class GeneralMoveGeneratorTest
     public void DiscoveredCheckAfterEnPassantCapture()
     {
         var originalPosition = new Position("8/8/8/k1pP3R/8/8/8/n4K2 w - c6 0 1");
-        var enPassantMove = MoveGenerator.GenerateAllMoves(originalPosition, new Move[Constants.MaxNumberOfPossibleMovesInAPosition]).Single(m => m.IsEnPassant());
+        Span<Move> moves = stackalloc Move[Constants.MaxNumberOfPossibleMovesInAPosition];
+        var enPassantMove = MoveGenerator.GenerateAllMoves(originalPosition, moves).ToArray().Single(m => m.IsEnPassant());
         var positionAferEnPassant = new Position(originalPosition, enPassantMove);
 
-        foreach (var move in MoveGenerator.GenerateAllMoves(positionAferEnPassant, new Move[Constants.MaxNumberOfPossibleMovesInAPosition]))
+        moves = stackalloc Move[Constants.MaxNumberOfPossibleMovesInAPosition];
+        foreach (var move in MoveGenerator.GenerateAllMoves(positionAferEnPassant, moves))
         {
             if (new Position(positionAferEnPassant, move).IsValid())
             {