--- a/tests/Lynx.Test/MoveGeneration/RegressionTest.cs
+++ b/tests/Lynx.Test/MoveGeneration/RegressionTest.cs
@@ -19,7 +19,8 @@ public class MoveGeneratorRegressionTest : BaseTest
         Assert.True(moves.Exists(m => m.PromotedPiece() != default && !m.IsCapture()));
         Assert.True(moves.Exists(m => m.IsDoublePawnPush()));
 
-        var captures = MoveGenerator.GenerateAllCaptures(position, new Move[Constants.MaxNumberOfPossibleMovesInAPosition]).ToList();
+        Span<Move> moveSpan = stackalloc Move[Constants.MaxNumberOfPossibleMovesInAPosition];
+        var captures = MoveGenerator.GenerateAllCaptures(position, moveSpan).ToArray().ToList();
 
         Assert.True(moves.Exists(m => m.IsShortCastle()));
         Assert.True(moves.Exists(m => m.IsLongCastle()));