--- a/tests/Lynx.Test/PerftTestSuite.cs
+++ b/tests/Lynx.Test/PerftTestSuite.cs
@@ -771,6 +771,6 @@ public class PerftTestSuite
 
     private static void Validate(string fen, int depth, long expectedNumberOfNodes)
     {
-        Assert.AreEqual(expectedNumberOfNodes, Perft.ResultsImpl(new Position(fen), depth, default, new Move[Constants.MaxNumberOfPossibleMovesInAPosition]));
+        Assert.AreEqual(expectedNumberOfNodes, Perft.ResultsImpl(new Position(fen), depth, default));
     }
 }